use marathon

+++ recipes/mesos-install.rb
@@ -1,6 +1,6 @@
 execute "install mesos" do
   user "root"
-  command <<EOF
+  command <<EOL
 apt-key adv --keyserver keyserver.ubuntu.com --recv E56151BF
 export DISTRO=$(lsb_release -is | tr '[:upper:]' '[:lower:]')
 export CODENAME=$(lsb_release -cs)
@@ -8,5 +8,6 @@ echo "deb http://repos.mesosphere.io/${DISTRO} ${CODENAME} main" | \
   sudo tee /etc/apt/sources.list.d/mesosphere.list
 apt-get update
 apt-get -y install mesos
-EOF
+EOL
+not_if "dpkg -s mesos"
 end
